# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GoWin Sports Betting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users, admin]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users, admin]
      summary: Get user by ID (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Wallet ────────────────────────────────────────────────────────────────
...
  /sports:
    get:
      operationId: listSports
      tags: [sports]
      summary: List all sports
      responses:
        "200":
          description: Sports list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sport"
    post:
      operationId: createSport
      tags: [sports, admin]
      summary: Create a sport (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sport"
  /leagues:
    get:
      operationId: listLeagues
      tags: [sports]
      summary: List leagues, optionally filtered by sport
      parameters:
        - name: sportId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Leagues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/League"
    post:
      operationId: createLeague
      tags: [sports, admin]
      summary: Create a league (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeagueInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
  /teams:
    get:
      operationId: listTeams
      tags: [sports]
      summary: List teams
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [sports, admin]
      summary: Create a team (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /fixtures:
    get:
      operationId: listFixtures
      tags: [sports]
      summary: List fixtures
      parameters:
        - name: leagueId
          in: query
          schema:
            type: integer
        - name: sportId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, live, finished]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Fixtures
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixtureListResponse"
    post:
      operationId: createFixture
      tags: [sports, admin]
      summary: Create a fixture (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FixtureInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
  /fixtures/{id}:
    get:
      operationId: getFixture
      tags: [sports]
      summary: Get fixture with markets and odds
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fixture detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixtureDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateFixture
      tags: [sports, admin]
      summary: Update fixture (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FixtureUpdate"
      responses:
        "200":
          description: Updated fixture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
...
  /markets:
    post:
      operationId: createMarket
      tags: [sports, admin]
      summary: Create a market for a fixture (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
  /odds/{id}:
    patch:
      operationId: updateOdds
      tags: [sports, admin]
      summary: Update odds value (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OddsUpdate"
      responses:
        "200":
          description: Updated odds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Odds"

  # ── Bets ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ──────────────────────────────────────────────────────────────
...
    UserDetail:
      type: object
      required: [id, username, email, role, createdAt, wallet]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        wallet:
          $ref: "#/components/schemas/Wallet"
    UserListResponse:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserDetail"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Wallet ────────────────────────────────────────────────────────────
    Wallet:
      type: object
      required: [id, userId, balance]
      properties:
        id:
          type: integer
        userId:
          type: integer
        balance:
          type: number
          format: double
...
    Sport:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: ["string", "null"]
    SportInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        icon:
          type: string
    League:
      type: object
      required: [id, sportId, name]
      properties:
        id:
          type: integer
        sportId:
          type: integer
        name:
          type: string
        sport:
          $ref: "#/components/schemas/Sport"
    LeagueInput:
      type: object
      required: [sportId, name]
      properties:
        sportId:
          type: integer
        name:
          type: string
    Team:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        logo:
          type: ["string", "null"]
    TeamInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        logo:
          type: string
    Fixture:
      type: object
      required: [id, leagueId, homeTeamId, awayTeamId, startTime, status]
      properties:
        id:
          type: integer
        leagueId:
          type: integer
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        startTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [upcoming, live, finished, cancelled]
        scoreHome:
          type: ["integer", "null"]
        scoreAway:
          type: ["integer", "null"]
        league:
          $ref: "#/components/schemas/League"
        homeTeam:
          $ref: "#/components/schemas/Team"
        awayTeam:
          $ref: "#/components/schemas/Team"
    FixtureDetail:
      type: object
      required: [id, leagueId, homeTeamId, awayTeamId, startTime, status, markets]
      properties:
        id:
          type: integer
        leagueId:
          type: integer
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        startTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [upcoming, live, finished, cancelled]
        scoreHome:
          type: ["integer", "null"]
        scoreAway:
          type: ["integer", "null"]
        league:
          $ref: "#/components/schemas/League"
        homeTeam:
          $ref: "#/components/schemas/Team"
        awayTeam:
          $ref: "#/components/schemas/Team"
        markets:
          type: array
          items:
            $ref: "#/components/schemas/Market"
    FixtureListResponse:
      type: object
      required: [fixtures, total, page, limit]
      properties:
        fixtures:
          type: array
          items:
            $ref: "#/components/schemas/Fixture"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    FixtureInput:
      type: object
      required: [leagueId, homeTeamId, awayTeamId, startTime]
      properties:
        leagueId:
          type: integer
        homeTeamId:
          type: integer
        awayTeamId:
          type: integer
        startTime:
          type: string
          format: date-time
    FixtureUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [upcoming, live, finished, cancelled]
        scoreHome:
          type: integer
        scoreAway:
          type: integer
        startTime:
          type: string
          format: date-time
...
    Odds:
      type: object
      required: [id, marketId, selection, oddsValue]
      properties:
        id:
          type: integer
        marketId:
          type: integer
        selection:
          type: string
        oddsValue:
          type: number
          format: double
    Market:
      type: object
      required: [id, fixtureId, marketType, odds]
      properties:
        id:
          type: integer
        fixtureId:
          type: integer
        marketType:
          type: string
        odds:
          type: array
          items:
            $ref: "#/components/schemas/Odds"
    MarketInput:
      type: object
      required: [fixtureId, marketType, selections]
      properties:
        fixtureId:
          type: integer
        marketType:
          type: string
        selections:
          type: array
          items:
            $ref: "#/components/schemas/OddsInput"
    OddsInput:
      type: object
      required: [selection, oddsValue]
      properties:
        selection:
          type: string
        oddsValue:
          type: number
          format: double
    OddsUpdate:
      type: object
      required: [oddsValue]
      properties:
        oddsValue:
          type: number
          format: double

    # ── Bets ──────────────────────────────────────────────────────────────
...