# lib/api-zod/src/generated/types/market.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GoWin Sports Betting Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Odds } from './odds';

export interface Market {
  id: number;
  fixtureId: number;
  marketType: string;
  odds: Odds[];
}