# lib/api-zod/src/generated/types/marketInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GoWin Sports Betting Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OddsInput } from './oddsInput';

export interface MarketInput {
  fixtureId: number;
  marketType: string;
  selections: OddsInput[];
}