# lib/api-zod/src/generated/types/oddsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GoWin Sports Betting Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OddsInput {
  selection: string;
  oddsValue: number;
}